added helper tests
