+++ js/__tests__/actions/ProfileAction.spec.ts
@@ -1,28 +1,28 @@
 import {
-    FETCH_PENDING,
-    FETCH_SUCCESS,
-    FETCH_FAIL
-} from '../src/constants';
-import { TodosAction } from '../src/actions';
+    PROFILE_PENDING,
+    PROFILE_SUCCESS,
+    PROFILE_FAIL
+} from '../../src/constants';
+import { ProfileAction } from '../../src/actions';
 
 const mockEmail = 'test@test.com';
 const mockPassword = 'test';
 
 const mockPending = {
-    type: FETCH_PENDING
+    type: PROFILE_PENDING
 };
 const mockFailed = {
-    type: FETCH_FAIL,
+    type: PROFILE_FAIL,
     payload: {
         status: 401,
         message: ''
     }
 };
 const mockSuccess = {
-    type: FETCH_SUCCESS,
+    type: PROFILE_SUCCESS,
     payload: {
         status: 200,
-        data: [],
+        data: {},
     }
 };
 const TOKEN = 'abc123';
@@ -31,7 +31,7 @@ const TOKEN = 'abc123';
 jest.mock('axios', () => ({
     get: jest.fn((url, opts) => {
         if (opts.headers.Authorization === 'Bearer abc123') {
-            return Promise.resolve({ status: 200, data: { todos: [], message: '' } });
+            return Promise.resolve({ status: 200, data: { profile: {} } });
         } else {
             return Promise.resolve({ status: 401, data: { message: '' }});
         }
@@ -41,18 +41,18 @@ jest.mock('axios', () => ({
 describe('Profile action', () => {
     test('It should return a function', () => {
         const dispatch = jest.fn();
-        expect(TodosAction(dispatch)).toBeInstanceOf(Function);
+        expect(ProfileAction(dispatch)).toBeInstanceOf(Function);
     });
     test('It should call dispatch', () => {
         const dispatch = jest.fn();
-        const thunk = TodosAction(dispatch);
+        const thunk = ProfileAction(dispatch);
         return thunk(TOKEN).then(() => {
             expect(dispatch.mock.calls.length).toBeGreaterThan(0);
         });
     })
     test('It should dispatch a success action', () => {
         const dispatch = jest.fn();
-        const thunk = TodosAction(dispatch);
+        const thunk = ProfileAction(dispatch);
         return thunk(TOKEN).then(() => {
             const [ pending, success ] = dispatch.mock.calls;
             expect(pending[0]).toEqual(mockPending);
@@ -61,7 +61,7 @@ describe('Profile action', () => {
     });
     test('It should dispatch a fail action', () => {
         const dispatch = jest.fn();
-        const thunk = TodosAction(dispatch);
+        const thunk = ProfileAction(dispatch);
         return thunk('wrong').then(() => {
             const [ pending, failed ] = dispatch.mock.calls;
             expect(pending[0]).toEqual(mockPending);
