+++ js/__tests__/actions/TodosAction.spec.ts
@@ -1,67 +1,59 @@
 import {
-    LOGIN_PENDING,
-    LOGIN_SUCCESS,
-    LOGIN_FAIL
-} from '../src/constants';
-import { LoginAction } from '../src/actions';
+    FETCH_PENDING,
+    FETCH_SUCCESS,
+    FETCH_FAIL
+} from '../../src/constants';
+import { TodosAction } from '../../src/actions';
 
 const mockEmail = 'test@test.com';
 const mockPassword = 'test';
 
 const mockPending = {
-    type: LOGIN_PENDING
+    type: FETCH_PENDING
 };
 const mockFailed = {
-    type: LOGIN_FAIL,
+    type: FETCH_FAIL,
     payload: {
         status: 401,
         message: ''
     }
 };
 const mockSuccess = {
-    type: LOGIN_SUCCESS,
+    type: FETCH_SUCCESS,
     payload: {
         status: 200,
-        data: {},
-        token: 'test'
+        data: [],
     }
 };
-
-const localStorageMock = {
-    getItem: jest.fn(),
-    setItem: jest.fn(),
-    clear: jest.fn(),
-};
-const myGlobal: any = global;
-myGlobal.localStorage = localStorageMock;
+const TOKEN = 'abc123';
 
 // mock axios
 jest.mock('axios', () => ({
-    get: jest.fn((url) => {
-        if (url.includes('email=test@test.com') && url.includes('password=test')) {
-            return Promise.resolve({ status: 200, data: { profile: {}, token: 'test' } });
+    get: jest.fn((url, opts) => {
+        if (opts.headers.Authorization === 'Bearer abc123') {
+            return Promise.resolve({ status: 200, data: { todos: [], message: '' } });
         } else {
-            return Promise.resolve({ status: 401, data: mockFailed.payload });
+            return Promise.resolve({ status: 401, data: { message: '' }});
         }
     })
 }));
 
-describe('Login action', () => {
+describe('Profile action', () => {
     test('It should return a function', () => {
         const dispatch = jest.fn();
-        expect(LoginAction(dispatch)).toBeInstanceOf(Function);
+        expect(TodosAction(dispatch)).toBeInstanceOf(Function);
     });
     test('It should call dispatch', () => {
         const dispatch = jest.fn();
-        const thunk = LoginAction(dispatch);
-        return thunk(mockEmail, mockPassword).then(() => {
+        const thunk = TodosAction(dispatch);
+        return thunk(TOKEN).then(() => {
             expect(dispatch.mock.calls.length).toBeGreaterThan(0);
         });
     })
     test('It should dispatch a success action', () => {
         const dispatch = jest.fn();
-        const thunk = LoginAction(dispatch);
-        return thunk(mockEmail, mockPassword).then(() => {
+        const thunk = TodosAction(dispatch);
+        return thunk(TOKEN).then(() => {
             const [ pending, success ] = dispatch.mock.calls;
             expect(pending[0]).toEqual(mockPending);
             expect(success[0]).toEqual(mockSuccess);
@@ -69,8 +61,8 @@ describe('Login action', () => {
     });
     test('It should dispatch a fail action', () => {
         const dispatch = jest.fn();
-        const thunk = LoginAction(dispatch);
-        return thunk('wrong', 'wrong').then(() => {
+        const thunk = TodosAction(dispatch);
+        return thunk('wrong').then(() => {
             const [ pending, failed ] = dispatch.mock.calls;
             expect(pending[0]).toEqual(mockPending);
             expect(failed[0]).toEqual(mockFailed);
